pyproject.toml
[tool.poetry]
name = "intakectf-2022"
version = "0.1.0"
description = ""
authors = ["Oshawk <oshawk@protonmail.com>"]
readme = "README.md"
packages = [{include = "intakectf_2022"}]

[tool.poetry.dependencies]
python = "^3.10"
pycryptodome = "^3.15.0"
Flask = "^2.2.2"
requests = "^2.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
